+++ resources/js/admin.js
@@ -1,7 +1,9 @@
 window.debounce = require('debounce');
 window.Vue = require('vue').default;
 
-Vue.component('linked-combos', require('./components/vue/linked-combos').default);
 Vue.component('admin-links-input', require('./components/admin/LinksInput.vue').default);
+Vue.component('autocomplete', require('./components/Autocomplete.vue').default);
+Vue.component('linked-combos', require('./components/vue/linked-combos').default);
+Vue.component('query-string', require('./components/QueryString.vue').default);
 
 new Vue({ el: '#wrapper' })