+++ resources/views/components/admin/checkbox.blade.php
@@ -1,0 +1,4 @@
+@props(['name'])
+
+<input type="hidden" name="{{ $name }}" value="0">
+<input type="checkbox" name="{{ $name }}" value="1" {{ $attributes }}>