+++ resources/views/components/item_author.blade.php
@@ -1,13 +1,13 @@
 {{-- comments kept for code indentation without whitespace --}}
-@if (is_string($author))
-    <a class="underline" href="{{ route('frontend.catalog.index', ['author' => $author]) }}">{{ formatName($author) }}</a>{{--
---}}@elseif ($author->pivot->role === 'autor/author')
+@empty ($author->authority)
+    <a class="underline" href="{{ route('frontend.catalog.index', ['author' => $author->name]) }}">{{ formatName($author->name) }}</a>{{--
+--}}@elseif ($author->authority->pivot->role === 'autor/author')
     <span itemprop="creator" itemscope itemtype="http://schema.org/Person">
-        <a class="underline" href="{{ $author->getUrl() }}" itemprop="sameAs">
-            <span itemprop="name">{{ formatName($author->name) }}</span>{{--
+        <a class="underline" href="{{ $author->authority->getUrl() }}" itemprop="sameAs">
+            <span itemprop="name">{{ formatName($author->authority->name) }}</span>{{--
         --}}</a>{{--
     --}}</span>{{--
 --}}@else
-    <a class="underline" href="{{ $author->getUrl() }}">{{ formatName($author->name) }}</a>
-    &ndash; {{ $author::formatMultiAttribute($author->pivot->role) }}{{--
+    <a class="underline" href="{{ $author->authority->getUrl() }}">{{ formatName($author->authority->name) }}</a>
+    &ndash; {{ $author->authority::formatMultiAttribute($author->authority->pivot->role) }}{{--
 --}}@endif