+++ resources/views/layouts/admin.blade.php
@@ -17,6 +17,7 @@
         <!-- CSS are placed here -->
         <script src="https://use.fontawesome.com/73587c90bb.js"></script>
         <link rel="stylesheet" href="//netdna.bootstrapcdn.com/bootstrap/3.1.1/css/bootstrap.min.css">
+        <link rel="stylesheet" type="text/css" href="{{ mix('/css/app-tailwind.css') }}" />
         {!! Html::style('css/sb-admin.css') !!}
         {!! Html::style('css/ladda-themeless.min.css') !!}
         {!! Html::style('css/bootstrap-wysihtml5.css') !!}
@@ -109,8 +110,11 @@
                         <li class="uppercase text-sm font-semibold text-muted pl-4 pt-4 border-b-0">
                             Homepage
                         </li>
-                        <li class="mb-4 border-b-0">
-                            <a href="{{ route('featured-pieces.index') }}"><i class="fa fa-image fa-fw"></i> Odporúčaný obsah</a>
+                        <li class="border-b-0">
+                            <a href="{{ route('featured-pieces.index') }}"><i class="fa fa-newspaper-o fa-fw"></i> Odporúčaný obsah</a>
+                        </li>
+                        <li class="border-b-0 mb-4">
+                            <a href="{{ route('featured-artworks.index') }}"><i class="fa fa-image fa-fw"></i> Vybrané diela</a>
                         </li>
                         @endcan
                         @can('administer')