+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 
+## [2.46.0] - 2022-02-25
+### Added
+- Featured Artworks back-end section
+- new Tailwind component for back-end/admin
+
+### Changed
+- font loading (fix for Source Serif Pro)
+- Tailwind scale to start from 1rem = 16px
+
 ## [2.45.0] - 2022-02-18
 ### Added
 - age group "all" and media type "activities" to edu article tags